requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
certifi==2024.6.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jupyter-client==8.6.2
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.34.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
zipp==3.19.2
